moving to configuration package
