+++ pkg/configuration/load_configuration.go
@@ -1,4 +1,4 @@
-package controller
+package configuration
 
 import (
 	"context"
