+++ pkg/configuration/load_configuration_test.go
@@ -1,4 +1,4 @@
-package controller
+package configuration
 
 import (
 	"context"
